Move path definitions forward in _bashrc
This way they are defined even for non-interactive sessions.  An
example where this is necessary is for the target of an
  hg push ssh://...

Where you can't install Mercurial systemwide on the target.  Mercurial
logs in using ssh without setting up a prompt, but we still need to
define a PATH and PYTHONPATH including the Mercurial install
locations:
  ~/bin/hg
  ~/lib/python2.5/site-packages/mercurial-1.5.3-py2.5-linux-i686.egg/
Actually, the .egg directory is added in
  ~/lib/python2.5/site-packages/easy-install.pth
So we only need
  ~/lib/python2.5/site-packages/
in our PYTHONPATH.