Re-organize dotfiles management scripts.
See `README` for details on the new framework, which I think is more
flexible and transparent.

Moved the dotfiles themselves from `_XXX` to `src/.XXX`, which makes
it easy to keep them separate from the controlling scripts, and avoids
the need to make string replacements in the filesnames.

Removed `Makefile` (in favor of calling the scripts management scripts
directly).

Added `.gitigore`, ignoring locally created directories.